refactor: move shared utilities into the components folder to fix build
