+++ packages/lib/components/select/select.ts
@@ -4,7 +4,7 @@ import { classMap } from 'lit/directives/class-map.js';
 
 import '../icon';
 
-import { FormControl } from '../../shared/formControl';
+import { FormControl } from '../shared/formControl';
 import type { Option } from './option';
 import type { OptionValue } from './types';
