+++ packages/lib/components/shared/inputModeDetector.ts
