+++ packages/lib/components/toast/toast-item.ts
@@ -4,7 +4,7 @@ import { classMap } from 'lit/directives/class-map.js';
 
 import a11yStyles from '../../global-css/a11y.css?inline';
 import animationStyles from '../../global-css/animations.css?inline';
-import { BreakpointObserver } from '../../shared/breakpointObserver';
+import { BreakpointObserver } from '../shared/breakpointObserver';
 import buttonStyles from '../button/button.css?inline';
 import { addIcons } from '../icon';
 import { type IconName, checkCircle, close, errorCircle, infoCircle, warning } from '../icon/iconRegistry';
